Divide Livenet into modules.
