+++ odra-casper/rpc-client/src/casper_client/configuration.rs
@@ -2,6 +2,7 @@ use crate::casper_client::{
     ENV_ACCOUNT_PREFIX, ENV_CHAIN_NAME, ENV_CSPR_CLOUD_AUTH_TOKEN, ENV_EVENTS_ADDRESS,
     ENV_LIVENET_ENV_FILE, ENV_NODE_ADDRESS, ENV_SECRET_KEY, ENV_TTL
 };
+use crate::log;
 use crate::utils::{get_env_variable, get_optional_env_variable};
 use casper_client::Verbosity;
 use casper_types::TimeDiff;
@@ -15,7 +16,6 @@ pub const DEFAULT_GAS_TOLERANCE: u8 = 5;
 pub struct CasperClientConfiguration {
     pub node_address: String,
     pub events_url: String,
-    pub rpc_id: String,
     pub chain_name: String,
     pub secret_keys: Vec<SecretKey>,
     pub secret_key_paths: Vec<String>,
@@ -36,6 +36,9 @@ impl CasperClientConfiguration {
         // Load .env
         dotenv::dotenv().ok();
 
+        // Initialize logging from environment variable
+        log::init_log_level();
+
         let node_address = get_env_variable(ENV_NODE_ADDRESS);
         let chain_name = get_env_variable(ENV_CHAIN_NAME);
         let events_url = get_env_variable(ENV_EVENTS_ADDRESS);
@@ -46,7 +49,6 @@ impl CasperClientConfiguration {
         let (secret_keys, secret_key_paths) = Self::secret_keys_from_env();
         CasperClientConfiguration {
             node_address,
-            rpc_id: "1".to_string(),
             chain_name,
             secret_keys,
             secret_key_paths,
